app: replace all number types in api with BigInteger

+++ app/app/src/main/java/de/stustapay/stustapay/repository/CashierRepository.kt
@@ -3,7 +3,7 @@ package de.stustapay.stustapay.repository
 import de.stustapay.api.models.CashRegister
 import de.stustapay.api.models.CashRegisterStocking
 import de.stustapay.api.models.UserInfo
-import de.stustapay.stustapay.model.UserTag
+import de.stustapay.api.models.UserTag
 import de.stustapay.stustapay.net.Response
 import de.stustapay.stustapay.netsource.CashierRemoteDataSource
 import javax.inject.Inject
@@ -15,7 +15,7 @@ class CashierRepository @Inject constructor(
         return cashierRemoteDataSource.getCashierStockings()
     }
 
-    suspend fun equipCashier(tagid: ULong, registerId: Int, stockingId: Int): Response<Unit> {
+    suspend fun equipCashier(tagid: ULong, registerId: ULong, stockingId: ULong): Response<Unit> {
         return cashierRemoteDataSource.equipCashier(tagid, registerId, stockingId)
     }
