+++ app/app/src/main/java/de/stustapay/stustapay/ui/sale/SaleViewModel.kt
@@ -3,8 +3,8 @@ package de.stustapay.stustapay.ui.sale
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
+import de.stustapay.api.models.UserTag
 import de.stustapay.stustapay.model.CompletedSale
-import de.stustapay.stustapay.model.UserTag
 import de.stustapay.stustapay.net.Response
 import de.stustapay.stustapay.repository.SaleRepository
 import de.stustapay.stustapay.repository.TerminalConfigRepository
@@ -255,9 +255,9 @@ class SaleViewModel @Inject constructor(
                             ready = true,
                             buttons = terminalConfig.config.till?.buttons?.associate {
                                 Pair(
-                                    it.id,
+                                    it.id.intValue(),
                                     SaleItemConfig(
-                                        id = it.id,
+                                        id = it.id.intValue(),
                                         caption = it.name,
                                         price = SaleItemPrice.fromTerminalButton(it),
                                         returnable = it.isReturnable,
